// src/model/send-solution.model.ts
export interface SendSolution {
    solution: string;
    employeeId: number;
    name: string;
    role: string;
    comments: string;
    score: number;
    doneMaze1: boolean;
    doneMaze2: boolean;
    doneMaze3: boolean;
    doneMaze4: boolean;
    doneMaze5: boolean;
    doneMaze6: boolean;
    doneMaze7: boolean;
    doneMaze8: boolean;
    doneMaze9: boolean;
    doneMaze10: boolean;
}